Update formatting and names
